+++ src/js/components/Rebus.js
@@ -2,6 +2,8 @@ import { createComponent } from '../mini';
 import { connect } from '../store';
 import { Word } from './Word';
 
+const rebusRef = localStorage.getItem('flagBR') === 'true' ? '?rebus-br=' : '?rebus=';
+
 export function Rebus(props, ...children) {
   return connect(
     createComponent({
@@ -17,13 +19,11 @@ export function Rebus(props, ...children) {
       },
       componentDidUpdate() {
         const rebus = this.props.rebuses[this.props.current];
-        const rebusRef =
-          (localStorage.getItem('flagBR') === 'true' ? '?rebus-br=' : '?rebus=') + rebus.id;
         /* If history API isn't available, we shouldn't revert to the more widely available `window.location.href`, 
         as it incurs a new HTTP request and thus results in an infinite loop (and breaks SPAs). */
         if (window.history) {
           // Adds 'rebus' query parameter to end of URL. Should be endpoint-agnostic.
-          window.history.pushState('', '', `${rebusRef}`);
+          window.history.pushState('', '', `${rebusRef + rebus.id}`);
         }
         if (rebus.isAnswered) {
           this.$parent.querySelector('.change-button--next').focus();
